requirements.txt
# ── Web framework ─────────────────────────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1

# ── Embeddings ────────────────────────────────────────────────────────────────
# all-MiniLM-L6-v2 ships inside sentence-transformers; torch is its runtime.
# Pinning CPU-only torch keeps the image ~1.5 GB smaller than the GPU variant.
sentence-transformers==3.0.1
torch==2.3.0+cpu ; sys_platform != "darwin"
torch==2.3.0     ; sys_platform == "darwin"
--extra-index-url https://download.pytorch.org/whl/cpu

# ── Vector store ──────────────────────────────────────────────────────────────
# faiss-cpu: pure CPU FAISS build – no CUDA driver required at runtime.
faiss-cpu==1.8.0

# ── Data & ML ─────────────────────────────────────────────────────────────────
scikit-learn==1.5.0
numpy==1.26.4

# ── Dataset ───────────────────────────────────────────────────────────────────
# Included via scikit-learn (sklearn.datasets.fetch_20newsgroups).
# No additional package needed.

# ── Analysis / notebook (optional – not installed in the runtime Docker image) ─
matplotlib==3.9.0
seaborn==0.13.2
jupyterlab==4.2.0
ipywidgets==8.1.3
